FIX: backwards compat passing for table model

+++ finetune/nn/crf.py
@@ -60,7 +60,7 @@ def batch_viterbi_decode(logits, transition_matrix):
 def sequence_decode(logits, transition_matrix, use_crf):
     if not use_crf:
         return tf.argmax(input=logits, axis=-1), tf.nn.softmax(logits, -1)
-    return batch_viterbi_decode(logits, tf.convert_to_tensor(transition_matrix))
+    return batch_viterbi_decode(logits, tf.convert_to_tensor(transition_matrix, dtype=logits.dtype))
 
 
 # Everything below here is basically verbatim from tf_addons - If we find someone is maintaining this then we should use that instead
